Linux 1.7 Open Source Gold release Signed-off-by: Li, Xun <xun.li@email.com>

+++ SampleCode/PowerTransition/Enclave/Enclave.cpp
@@ -158,7 +158,7 @@ int increase_and_seal_data(size_t tid, struct sealed_buf_t* sealed_buf)
     // Ocall to print the unsealed secret data outside.
     // In theory, the secret data(s) SHOULD NOT be transferred outside the enclave as clear text(s).
     // So please DO NOT print any secret outside. Here printing the secret data to outside is only for demo.
-    snprintf(string_buf, BUFSIZ, "Thread %#x>: %d\n", (unsigned int)tid, temp_secret);
+    snprintf(string_buf, BUFSIZ, "Thread %#x>: %u\n", (unsigned int)tid, (unsigned int)temp_secret);
     print(string_buf);
     return 0;
 }
